requirements.txt
certifi==2020.6.20
charset-normalizer==2.0.12
decouple==0.0.7
Flask
idna==3.3
itsdangerous
Jinja2 
MarkupSafe
mutagen==1.45.1
pycryptodomex==3.14.1
python-decouple==3.6
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
spotipy==2.19.0
urllib3==1.26.9
websockets==10.2
Werkzeug
youtube-search==2.1.0
yt-dlp==2022.3.8.2
